Resolve side effect caused by previous commit
It seems in plugin.py the decoder was set to normal via the old style
and that causes a crash from the plugin because etDecoderMode could not
be imported anymore. As this is now done via a 'shadow' config and this
is already set correctly at a restart by enigma2 this needs not to be
set to 'normal' via autostart anymore. So remove this autostart code.